+++ skrl/agents/torch/ddpg/ddpg_rnn.py
@@ -14,6 +14,7 @@ from skrl.agents.torch import Agent
 from skrl.memories.torch import Memory
 from skrl.models.torch import Model
 from skrl.utils import ScopedTimer
+from skrl.utils.spaces.torch import compute_space_limits
 
 from .ddpg_cfg import DDPG_CFG
 
@@ -183,9 +184,9 @@ class DDPG_RNN(Agent):
             self._rnn_initial_states["policy"].append(torch.zeros(size, dtype=torch.float32, device=self.device))
 
         # clip noise bounds
-        if self.action_space is not None:
-            self.clip_actions_min = torch.tensor(self.action_space.low, device=self.device)
-            self.clip_actions_max = torch.tensor(self.action_space.high, device=self.device)
+        self._min_actions, self._max_actions = compute_space_limits(
+            self.action_space, device=self.device, none_if_unbounded="any"
+        )
 
     def act(
         self, observations: torch.Tensor, states: torch.Tensor | None, *, timestep: int, timesteps: int
@@ -223,7 +224,7 @@ class DDPG_RNN(Agent):
             if self.cfg.exploration_scheduler:
                 noises.mul_(self.cfg.exploration_scheduler(timestep, timesteps))
             actions.add_(noises)
-            actions.clamp_(min=self.clip_actions_min, max=self.clip_actions_max)
+            actions.clamp_(min=self._min_actions, max=self._max_actions)
 
             self.track_data("Exploration / Exploration noise (max)", torch.max(noises).item())
             self.track_data("Exploration / Exploration noise (min)", torch.min(noises).item())